fix(backend): move Express Request augmentation to ambient .d.ts
The Express Request/User type augmentation lived inside App.ts as a
regular module. ts-node only picks up augmentations from files in the
entry point's import graph, so running knex seed/migrate (which never
imports App.ts) failed to typecheck winston.ts after it started reading
req.account in #23250. Moving the declare global block to
src/@types/express.d.ts makes it ambient and ensures it loads
regardless of entry point.
