+++ packages/backend/tsconfig.json
@@ -22,7 +22,7 @@
         "skipLibCheck": true,
         "types": ["node", "jest", "passport", "express-session"]
     },
-    "files": ["src/@types/express-session.d.ts"],
+    "files": ["src/@types/express-session.d.ts", "src/@types/express.d.ts"],
     "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.json"],
     "exclude": ["dist", "node_modules", "src/ee/services/McpService/mcp-chart-app"],
     "references": [
